+++ LogicBuilder.Expressions.Utils/Expansions/Expansion.cs
@@ -5,9 +5,9 @@ namespace LogicBuilder.Expressions.Utils.Expansions
 {
     abstract public class Expansion(string memberName, Type memberType, Type parentType, List<string> selects)
     {
-        public string MemberName { get; set; } = memberName;
-        public Type MemberType { get; set; } = memberType;
-        public Type ParentType { get; set; } = parentType;
-        public List<string> Selects { get; set; } = selects;
+        public string MemberName { get; } = memberName;
+        public Type MemberType { get; } = memberType;
+        public Type ParentType { get; } = parentType;
+        public List<string> Selects { get; } = selects;
     }
 }